fix: preserve complex data in functions that used real-dtype arrays
Several functions allocated real-dtype arrays (or cast to float) and
silently dropped imaginary parts:

- optimal_clone: q_a was np.zeros((dim, dim)) (real), truncating complex
  contributions; build it with dtype=complex.
- ghz: the state vector was real, dropping complex coefficients; take the
  dtype from the coefficients (as w_state now does).
- mutual_coherence: cast inputs to float, discarding imaginary parts (and
  making the later np.conj a no-op); use a complex dtype and add a guard
  against zero vectors, which previously produced inf/nan.
- is_k_incoherent: the purity test used trace(mat @ mat); use
  trace(mat @ mat.conj().T) so it stays correct (and real) for complex
  inputs. For a Hermitian density matrix this is unchanged.

Add tests for complex ghz coefficients, complex mutual_coherence vectors,
and the zero-vector guard.

+++ toqito/states/ghz.py
@@ -74,8 +74,8 @@ def ghz(dim: int, num_qubits: int, coeff: list[int] | None = None) -> np.ndarray
     if not np.isclose(norm, 1.0):
         coeff = coeff / norm
 
-    # Initialize the GHZ state vector.
-    ghz_state = np.zeros((dim**num_qubits, 1))
+    # Initialize the GHZ state vector, matching the coefficient dtype so complex coefficients are preserved.
+    ghz_state = np.zeros((dim**num_qubits, 1), dtype=coeff.dtype)
     # Fill the state vector with the corresponding coefficients.
     for i in range(dim):
         # Calculate the index for the tensor product state |i, i, ..., i>.
